Automatically decided when to remake ninja.build
Ninja's config file (build.ninja) needs to be rebuilt whenever a change
is made to one of the *.settings files in rosetta, which collectively
control which files are part of the build.  This commits adds some code
to compare the modification times of those files to the modification
time build.ninja, and rebuilds it if necessary.  This commit also
removes the --update option, which was previously provided so the user
could manually trigger this process.