Fix a bug in the code to run specific tests.
This feature still doesn't work, because the core_init() that's used in
unit test stupidly complains about the extra command-line argument.
This is a bug in rosetta, but as of yet I haven't thought of a way to
work around it.

One suboptimal option is to find a way to work around the problem in
C++, presumably by calling core_init_with_argv() or something like that.
Then I could change the boilerplate script to include that work-around,
and that would at least cover the common cases.